add SafetyImpactSelectorView
